requirements.txt
Flask==3.0.0
google-cloud-bigquery==3.13.0
